requirements.txt
streamlit==1.32.0
langchain==0.2.0
langchain-community==0.2.0
langchain-google-genai==0.1.0
google-generativeai==0.8.0
python-dotenv==1.0.0
pydantic==2.6.0
pandas==2.2.0
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.18.0
numpy==1.26.0
scikit-learn==1.4.0
openpyxl==3.1.2
xlrd==2.0.1
chardet==5.2.0
